+++ test/constants.ts
@@ -13,8 +13,39 @@ const IS_ARRAY:number[][]             = [[0, 1, 2], new Array(1, 2, 3), new Arra
 const IS_OBJECT:Record<string,any>[]  = [{bar:'foo'}, new Object(), Object.create(null), Object.create([])];
 const IS_FUNCTION:any[]               = [testFunction, testArrowFunction];
 const IS_NULLABLE:unknown[]           = [NaN, undefined, null];
+const IS_FILE:File[]                  = [new File(['file content'], 'file.txt', { type: 'text/plain' })];
+const IS_BLOB:Blob[]                  = [new Blob(new Array(100).fill('a'))];
+
+const NOT_BLOB:any[] = [
+    ...IS_NUMERIC,
+    ...IS_INTEGER,
+    ...IS_BOOLEAN,
+    ...IS_STRING,
+    ...IS_REGEXP,
+    ...IS_DATE,
+    ...IS_ARRAY,
+    ...IS_OBJECT,
+    ...IS_FUNCTION,
+    ...IS_NULLABLE,
+];
+
+const NOT_FILE:any[] = [
+    ...IS_NUMERIC,
+    ...IS_INTEGER,
+    ...IS_BOOLEAN,
+    ...IS_STRING,
+    ...IS_REGEXP,
+    ...IS_DATE,
+    ...IS_ARRAY,
+    ...IS_OBJECT,
+    ...IS_FUNCTION,
+    ...IS_NULLABLE,
+    ...IS_BLOB,
+];
 
 const NOT_ARRAY:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -26,6 +57,8 @@ const NOT_ARRAY:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_NUMERIC:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_BOOLEAN,
     ...IS_STRING,
     ...IS_REGEXP,
@@ -38,6 +71,8 @@ const NOT_NUMERIC:any[] = [
     new Number(Number.EPSILON),
 ];
 const NOT_STRING:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -49,6 +84,8 @@ const NOT_STRING:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_OBJECT:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -60,6 +97,8 @@ const NOT_OBJECT:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_DATE:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -71,6 +110,8 @@ const NOT_DATE:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_FORM_DATA:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -83,6 +124,8 @@ const NOT_FORM_DATA:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_FUNCTION:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -94,6 +137,8 @@ const NOT_FUNCTION:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_BOOLEAN:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_STRING,
@@ -105,6 +150,8 @@ const NOT_BOOLEAN:any[] = [
     ...IS_NULLABLE,
 ];
 const NOT_REGEXP:any[] = [
+    ...IS_FILE,
+    ...IS_BLOB,
     ...IS_NUMERIC,
     ...IS_INTEGER,
     ...IS_BOOLEAN,
@@ -139,11 +186,13 @@ const NOT_DATE_STRING:any[] = [
 ];
 
 const CONSTANTS = {
+    NOT_BLOB,
     NOT_ARRAY,
     NOT_ARRAY_WITH_EMPTY: [...NOT_ARRAY, []],
     NOT_BOOLEAN,
     NOT_DATE,
     NOT_DATE_STRING,
+    NOT_FILE,
     NOT_FORM_DATA,
     NOT_FUNCTION,
     NOT_INTEGER: [...NOT_NUMERIC, -5.2, 4.6, -150.3, Math.PI, new Number(1.12345), new Number(Number.EPSILON)],
@@ -159,6 +208,8 @@ const CONSTANTS = {
     IS_STRING,
     IS_REGEXP,
     IS_DATE,
+    IS_BLOB,
+    IS_FILE,
     IS_ARRAY,
     IS_OBJECT,
     IS_FUNCTION,